requirement.txt
aioredis==1.2.0
asgi-redis==1.4.3
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
boto3==1.9.164
botocore==1.12.164
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.6.1
daphne==2.3.0
Django==2.2.1
django-appconf==1.0.3
django-compressor==2.3
django-storages==1.7.1
Django-Template-Preprocess==1.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.14
drfdocs==0.0.11
ecdsa==0.13.2
future==0.17.1
hiredis==1.0.0
htmlmin==0.1.12
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.17
incremental==17.5.0
jmespath==0.9.4
more-itertools==7.0.0
msgpack==0.6.1
msgpack-python==0.5.6
numpy==1.16.3
packaging==19.0
Pillow==6.0.0
pluggy==0.12.0
psycopg2==2.8.2
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyparsing==2.4.0
pypi==2.1
pytest==4.6.2
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
python-env==1.0.0
python-jose==3.0.1
pytz==2019.1
rcssmin==1.0.6
redis==2.10.6
rjsmin==1.1.0
rsa==4.0
s3transfer==0.2.1
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
urllib3==1.25.3
virtualenv==16.6.0
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.5.1
zope.interface==4.6.0
